+++ docs/CHANGELOG.md
@@ -20,8 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Note in regex JSDoc that lookahead support is positive only
   - Clarify in the benchmarking README that `--experimental-strip-types` is a Node thing
   - Update a `NOTE` in the main README to be a `CAUTION` and move under the example
-
 - No longer exporting internal use only types.  Not publicly documented, so not considering this a breaking change
+- Ensure the `BufferedIndexOfAnchoredSearchState` benchmark comparison strategy properly resets its state
 
 ## [1.0.0] - 2026-01-23